Add account name to sidebar head.
When managing multiple accounts it's hard to determine which one you're
currently viewing. This patch makes sidebar look like this:

+++++++++++++++++++++++++
luke@lukesmith          +
----------------------- +
INBOX                 99+
Sent Mail               +
Drafts                  +
Trash                   +
Spam			+
+++++++++++++++++++++++++

The only problem with it is the necessity of changing sidebar_format, as with
its current value these 2 excess mailboxes are being appended with 0. That
leads to situation when size of an empty mailbox is being displayed as empty
string instead of 0. I consider it's not a big deal.