modify the `find` call in `mailsync` on Mac
The MacOS `find` command seems to be case insensitive. So the current
`mailsync` script results in counting each local inbox folder thrice,
resulting in the wrong # of new mails the user is notified about.
The commit implements another `test` command for `$(uname)` to decide
which arguments should be passed to `find`.

_In ref. to :_
https://github.com/LukeSmithxyz/mutt-wizard/issues/157#issuecomment-494132981