Fix script removal
The `script` variable you were referring to previously is actually the
relative path to a script, for example `bin/mw`.

This fails to remove the script, because, obviously, the path is
incorrect.

I'd suggest renaming `script` to `scriptPath` in line 13
and `shared` to `sharedPath` in line 18
(or whatever variable names you prefer that make it clearer)
to avoid confusion in the future.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>