Pinentry-tty and curses support
I did edit the line 65 and added directly "mbsync -a" instead of "mailsync...". In current state mutt-wizard doesn't work properly in CLI only environment without X started. This is first step to be able atleast mbsync with terminal only pinentry. The problem still remains in sending an email with terminal only pinentry. Temporary solution for is to activate gpg key before sending an email and send the email before the key times out.