changed ping address to a more friendly location (#399)
* changed ping address to a more friendly location 

`1.1.1.1` is not a friendly address for users from some parts of the world (e.g., China). Changed to `example.org` which should be more friendly to access.

* changed internet check by fallback

This is probably better since for most people it does the same thing and the first back up is Cloudfare's other DNS address which surprisingly can be accessed in more places (e.g., China) now. And changed the finall fallback as example.org.