Add option to disable SSL/STARTTLS for manual provider settings
There are scenarios where setting up SSL for a mail service would not
provide any security benefit, just complicate things or would be
outright impossible. For example, the unofficial emersion/hydroxide
Protonmail bridge is usually ran on your own computer or a home server,
thus it does not bother with implementing encryption. One might also
have a mail server on LAN where getting a cert is difficult, and is
only accessible to trusted machines.

In the manual settings, a prompt is added where the desired enctyption
method can be chosen. For regular users, sensible defaults are provided
both for encryption and ports that can be selected by pressing Enter.

Room for improvement: it's currently assumed that if SMTP has no TLS
support, the PLAIN authentication method is used. This is written out to
`msmtprc`, because by default it disallows sending password data in
cleartext, and fails if the server does not support GSSAPI. Maybe prompt
the user?

This might also be useful for the official Protonmail bridge, but I do
not know how that handles TLS.