:bug: Set protocol to `imap` when using ProtonMail
Part 1 of 2 for "Log-on not successful" bug when using ProtonMail.

The curl login check defaults to `imaps` when `protocol` isn't set.
Because ProtonMail isn't using any TLS, we have to use the `imap`
protocol.
The `protocol` variable was never set, so now we set it when the
ProtonMail flag (-p) is set.

Bug mentioned in:
Fixes:
- https://github.com/LukeSmithxyz/mutt-wizard/issues/608
- https://github.com/LukeSmithxyz/mutt-wizard/issues/622