:bug: Supply imap port to curl login check
Part 2 of 2 for "Log-on not successful" bug when using ProtonMail.

The curl login check uses the imap(s) protocol, which means it will use
the default ports. However, when using ProtonMail, these ports are
different (namely 1143 instead of 993).

We now use the port variable in the curl login command, defaulting to
993 when the port was not overridden.

The port variable was already set when the ProtonMail flag (-p) was set.

Bug mentioned in:
Fixes:
- https://github.com/LukeSmithxyz/mutt-wizard/issues/608
- https://github.com/LukeSmithxyz/mutt-wizard/issues/622