Make the indicator respect the attributes of the item below it.
Enabling cursor_overlay makes the indicator respect the attributes of the item below it. For example, you have made the status bar show the mailboxes with new emails, in bold, and your index to show the new emails in reverse. Having set cursor_overlay = yes keeps those attributes even when the indicator is over the mailbox or email in question.