Fix escaping in message IDs from notmuch
`+` was properly escaped but only once. `$` was not escaped at all.
This would result in messages with message IDs containing more than one
`+` or at least on `$` to not be shown after limiting.
`$` itself has to be escaped presumably for the shell. This results in
three prepended backslashes in the limit string and six in the config
file, as there backslashes themselves have to be escaped.